MUDANCAS NO BANCO (DOUBLE) E NOMES

+++ banco.sql
@@ -11,11 +11,11 @@ CREATE TABLE bem(
 	id 				SERIAL PRIMARY KEY,
 	nome 			VARCHAR(60) NOT NULL,
 	dt_adiquicao 	DATE NOT NULL,
-	valor_compra 	FLOAT NOT NULL,
+	valor_compra 	DOUBLE NOT NULL,
     turno 			INT NOT NULL,
 	vida_util 		DOUBLE NOT NULL,
 	dt_venda 		DATE,
-	valor_venda 	FLOAT,
+	valor_venda 	DOUBLE,
 	usado 			DOUBLE,
 	valor_residual 	FLOAT,
 	id_usuario  	INT NOT NULL REFERENCES usuario(id)
